Update patch set 1
Patch Set 1:

Hi Harman,

just looked at the patch, and the first thing i noticed is that you are including an Activator for the provider and consumer. Now afaik latest MDSAL requires you to do consumer and provider tracking via config-subsystem. So have you thought not to provide the activator way and to actually go via the config-subsystem? I myself not even clear why not to go via service publishing in OSGi but i'm certain that creating instances of objects that produce and consume MDSAL services expose it to race conditions, unless properly tracked.

Thanks,
Giovanni

Patch-set: 1