Update patch set 1
Patch Set 1:

(1 comment)

Harman,

while normally that would be ok, consider these archetype will be used to template applications. Now the approach in here to instantiate in the activator the business logic, especially for the consumer without worrying to track if the provider is there or not even worrying if the service provided is there or no would lead to ill-functioning applications. So please at least consider to write some sort of logic that assume that the service consumed may not be there. So folks would be warned of it.

my 2 cents,
Giovanni

Patch-set: 1
Label: Code-Review=0