Update patch set 2
Patch Set 2:

Hi Giovanni

I could not figure out a way where MDSAL gives me an information that a default provider is present or not. consumer throws an exception that default provider is not available if provider package is not present. I'll try to update code with config sub system to make sure they are loaded in sequence.

Patch-set: 2
Label: Verified=0