Bug 1264 - Multiple Usability Changes to MD-SAL App Archetype
1) Removed from the generate bundle the geration of any file under yang-gen-*.
These files are auto-generated when you build the yang file and thus not necessary.
2) Split the initial xml into a provider and consumer initial xml files.
Used naming (05 for provider, 06 for consumer) to show priority
3) Modified the generate to place the config subsystem xml files for producer / consumer
under their respective directories. (This way if they want to ship them, but not
auto deploy we have provided a best practice)
4) Added a file ma,ed Archetype_Next_Steps.README in the root folder which outlines
some quick next steps that people can follow in place of the wiki (it still references
wiki for more details)
5) Added comments to the .git ignore

Change-Id: Ica9c1ab9d377ed972a76b0fa081f8f1fa2af1564
Signed-off-by: Devin Avery <devin.avery@brocade.com>