Update patch set 3
Patch Set 3:

Not sure I followed your question completely, but any file under yang-gen-* is deleted and then regenerated by a 'mvn clean install' in the project. The files in those directories are generated based on the yang model and should not be modified by developers (so no need for us to auto generate them with the generator).

The <App>Module and <App>ModuleFactory are generated by the 'mvn clean install' if and only if they don't previously exist. So, since we want to add some intelligence into the initial <App>Module we still auto generate that, but I we are not adding anything to <App>ModuleFactory, so I removed that to let it be generated the first time you run 'mvn clean install'.

Does that answer your question?

As a side note, I found a small issue with this so I need to upload a new patch.

Patch-set: 3
Label: Verified=0